refactor: move post-processing to asciidoc lifecycle
