+++ docs/src/main/java/io/jooby/adoc/DocGenerator.java
@@ -7,11 +7,9 @@ package io.jooby.adoc;
 
 import static io.jooby.SneakyThrows.throwingConsumer;
 import static io.jooby.SneakyThrows.throwingFunction;
-import static java.util.function.Predicate.not;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -20,10 +18,8 @@ import java.time.Instant;
 import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.UUID;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Stream;
 
 import org.apache.commons.io.FileUtils;
@@ -35,7 +31,6 @@ import org.asciidoctor.Placement;
 import org.asciidoctor.SafeMode;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
-import org.jsoup.nodes.Element;
 
 import me.tongfei.progressbar.ProgressBarBuilder;
 import me.tongfei.progressbar.ProgressBarStyle;
@@ -52,9 +47,9 @@ public class DocGenerator {
 
     Path asciidoc = basedir.resolve("asciidoc");
 
-    /**
-     * Tree dir. The .adoc file became a directory
-     * modules/hikari.adoc => modules/hikari/index.html
+    /*
+      Tree dir. The .adoc file became a directory
+      modules/hikari.adoc => modules/hikari/index.html
      */
     String[] treeDirs = {"modules", "packaging", "usage", "migration"};
 
@@ -62,7 +57,7 @@ public class DocGenerator {
         Stream.of(treeDirs)
             .map(throwingFunction(dir -> countAdoc(asciidoc.resolve(dir))))
             .reduce(1, Integer::sum);
-    int steps = 7 + (doAscii ? adocCount : 0);
+    int steps = 6 + (doAscii ? adocCount : 0);
 
     ProgressBarBuilder pbb =
         new ProgressBarBuilder()
@@ -78,11 +73,11 @@ public class DocGenerator {
       }
       pb.step();
 
-      /** Wipe out directory: */
+      /* Wipe out directory: */
       FileUtils.cleanDirectory(outdir.toFile());
       pb.step();
 
-      /** Copy /images and /js: */
+      /* Copy /images and /js: */
       copyFile(
           outdir,
           // images
@@ -99,7 +94,6 @@ public class DocGenerator {
             createOptions(asciidoc, outdir, version, null));
         pb.step();
 
-        AtomicInteger m = new AtomicInteger();
         Stream.of(treeDirs)
             .forEach(
                 throwingConsumer(
@@ -112,20 +106,9 @@ public class DocGenerator {
                               module -> {
                                 processModule(asciidoctor, asciidoc, module, outdir, name, version);
                                 pb.step();
-                                m.incrementAndGet();
                               });
                     }));
       }
-
-      // post process
-      Files.walk(outdir)
-          .filter(it -> it.getFileName().toString().endsWith("index.html"))
-          .forEach(
-              throwingConsumer(
-                  it -> {
-                    Files.write(it, document(it).getBytes(StandardCharsets.UTF_8));
-                  }));
-      pb.step();
 
       // LICENSE
       Files.copy(
@@ -148,7 +131,7 @@ public class DocGenerator {
         Git git = new Git("jooby-project", "jooby.io", website);
         git.clone();
 
-        /** Clean: */
+        /* Clean: */
         FileUtils.deleteDirectory(website.resolve("images").toFile());
         FileUtils.deleteDirectory(website.resolve("js").toFile());
         FileUtils.deleteQuietly(website.resolve("index.html").toFile());
@@ -238,11 +221,10 @@ public class DocGenerator {
       indexlike = indexlike.resolve("index.html");
       Files.createDirectories(indexlike.getParent());
       Files.move(output, indexlike);
-      String content =
-          new String(Files.readAllBytes(indexlike), StandardCharsets.UTF_8)
+      String content = Files.readString(indexlike)
               .replace("js/", "../../js/")
               .replace("images/", "../../images/");
-      Files.write(indexlike, content.getBytes(StandardCharsets.UTF_8));
+      Files.writeString(indexlike, content);
     } catch (IOException x) {
       throw new IllegalStateException(x);
     }
@@ -318,120 +300,6 @@ public class DocGenerator {
       }
     }
     return name.toString();
-  }
-
-  private static String document(Path index) {
-    try {
-      Document doc = Jsoup.parse(index.toFile(), "UTF-8");
-      headerIds(doc);
-      languageTab(doc);
-      clipboard(doc);
-      externalLink(doc);
-      Document.OutputSettings settings = new Document.OutputSettings();
-      settings.prettyPrint(false);
-      settings.indentAmount(0);
-      settings.outline(false);
-      return doc.outputSettings(settings).toString();
-    } catch (NullPointerException x) {
-      throw new IllegalStateException("File: " + index, x);
-    } catch (IOException x) {
-      throw new IllegalStateException("File: " + index, x);
-    }
-  }
-
-  private static void externalLink(Document doc) {
-    for (Element a : doc.select("a")) {
-      String href = a.attr("href");
-      if (href.startsWith("http://") || href.startsWith("https://")) {
-        a.attr("target", "_blank");
-      }
-    }
-  }
-
-  private static void languageTab(Document doc) {
-    for (Element primary : doc.select(".listingblock.primary")) {
-      Element secondary = primary.nextElementSibling();
-      String secondaryTitle = secondary.selectFirst(".title").text().trim();
-      Element primaryContent = primary.selectFirst(".content");
-      Element secondaryContent = secondary.selectFirst(".content");
-      secondary.remove();
-      secondaryContent.remove();
-
-      Element title = primary.selectFirst(".title");
-
-      Element tabs = doc.createElement("div").attr("class", "switch");
-      Element tab1 = tabs.appendElement("div");
-      tab1.attr("class", "switch--item option-1 selected");
-      if (secondaryTitle.equalsIgnoreCase("Kotlin")) {
-        tab1.text("Java");
-      } else {
-        tab1.text(title.text());
-      }
-
-      if (title.text().trim().equalsIgnoreCase(tab1.text().trim())) {
-        title.remove();
-      }
-
-      Element tab2 = tabs.appendElement("div");
-      tab2.attr("class", "switch--item option-2");
-      tab2.text(secondaryTitle);
-      tabs.appendTo(primary);
-      primaryContent.addClass("option-1");
-      primaryContent.appendTo(primary);
-      secondaryContent.appendTo(primary);
-      secondaryContent.addClass("hidden").addClass("option-2");
-    }
-  }
-
-  private static void headerIds(Document doc) {
-    headerIds(doc, 2);
-    headerIds(doc, 3);
-    headerIds(doc, 4);
-    headerIds(doc, 5);
-  }
-
-  private static void headerIds(Document doc, int level) {
-    doc.select("h" + level).stream()
-        .filter(not(DocGenerator::isDiscrete))
-        .forEach(h -> {
-            String id = h.attr("id");
-            LinkedHashSet<String> name = new LinkedHashSet<>();
-            int parent = level - 1;
-            Element p = h.parents().select("h" + parent).first();
-            if (p != null && !isDiscrete(p)) {
-              String parentId = p.attr("id");
-              if (!parentId.isEmpty()) {
-                name.add(parentId);
-              }
-            }
-            name.add(id.replaceAll("([a-zA-Z0-9-]+)-\\d+$", "$1"));
-            String newId = String.join("-", name);
-            if (!id.equals(newId)) {
-              h.attr("id", newId);
-              h.select("a").stream()
-                  .filter(a -> a.attr("href").equals("#" + id) && !a.attr("class").isEmpty())
-                  .forEach(a -> a.attr("href", "#" + newId));
-            }
-          });
-  }
-
-  private static boolean isDiscrete(Element p) {
-    return p.hasClass("discrete");
-  }
-
-  private static void clipboard(Document doc) {
-    for (Element code : doc.select("code.hljs")) {
-      String id = "x" + Long.toHexString(UUID.randomUUID().getMostSignificantBits());
-      code.attr("id", id);
-      Element button = code.parent().appendElement("button");
-      button.addClass("clipboard");
-      button.attr("data-clipboard-target", "#" + id);
-      Element img = button.appendElement("img");
-      img.attr("src", "/images/clippy.svg");
-      img.attr("class", "clippy");
-      img.attr("width", "13");
-      img.attr("alt", "Copy to clipboard");
-    }
   }
 
   public static Path basedir() {
