+++ docs/src/main/java/io/jooby/adoc/JoobyExtensionRegistry.java
@@ -13,5 +13,6 @@ public class JoobyExtensionRegistry implements ExtensionRegistry {
   public void register(Asciidoctor asciidoctor) {
     asciidoctor.javaExtensionRegistry().block("dependency", DependencyProcessor.class);
     asciidoctor.javaExtensionRegistry().inlineMacro("javadoc", JavadocProcessor.class);
+    asciidoctor.javaExtensionRegistry().postprocessor(DocPostprocessor.class);
   }
 }
